extract MessageFeed from ChatPage
